rename modaldialog to dialog
