--- packages/ow-ui/src/components/ModalDialog/index.tsx
@@ -1,2 +1,0 @@
-export { ModalDialog } from "./ModalDialog"
-export type { ModalDialogProps } from "./ModalDialog"
