+++ packages/ow-ui/src/index.ts
@@ -3,7 +3,7 @@ export * from "./components/Badge"
 export * from "./components/Button"
 export * from "./components/Card"
 export * from "./components/Checkbox"
-export * from "./components/ModalDialog"
+export * from "./components/Dialog"
 export * from "./components/Textarea"
 export * from "./components/Toggle"
 export * from "./components/Typography"
